# Conflicts:
#	backend/src/main/java/io/diveni/backend/model/Session.java
#	backend/src/main/java/io/diveni/backend/service/projectmanagementproviders/jiracloud/JiraCloudService.java
#	backend/src/main/java/io/diveni/backend/service/projectmanagementproviders/jiraserver/JiraServerService.java
#	docs/.vitepress/config.js
#	docs/guide/develop.md
#	frontend/src/components/IssueTrackerComponent.vue
#	frontend/src/components/SignInWithJiraButtonComponent.vue
#	frontend/src/components/SignInWithJiraCloudButtonComponent.vue
#	frontend/src/components/SignInWithJiraServerButtonComponent.vue
#	frontend/src/locales/es.json
#	frontend/src/locales/it.json
#	frontend/src/views/PrepareSessionPage.vue

+++ frontend/src/views/SessionPage.vue
@@ -10,14 +10,40 @@
           }}
         </h1>
       </b-col>
-      <b-col cols="auto" >
+      <b-col>
+        <b-button
+          v-if="!autoReveal && !planningStart"
+          class="mr-3 autoRevealButtons optionButton"
+          variant="outline-dark"
+          @click="
+            autoReveal = true;
+            $event.target.blur();
+          "
+        >
+          <b-icon-eye-slash-fill class="bIcons" />
+          {{ $t("page.session.during.estimation.buttons.autoRevealOff") }}
+        </b-button>
+        <b-button
+          v-if="autoReveal && !planningStart"
+          class="mr-3 autoRevealButtons optionButton"
+          variant="outline-dark"
+          @click="
+            autoReveal = false;
+            $event.target.blur();
+          "
+        >
+          <b-icon-eye-fill class="bIcons" />
+          {{ $t("page.session.during.estimation.buttons.autoRevealOn") }}
+        </b-button>
+      </b-col>
+      <b-col cols="auto" class="mr-auto">
         <copy-session-id-popup
           v-if="planningStart"
           class="float-end"
           :session-id="session_sessionID"
         />
       </b-col>
-      <b-col cols="auto" id="sessionCloseCol">
+      <b-col id="sessionCloseCol" cols="auto">
         <session-close-button
           :is-planning-start="planningStart"
           :user-story-mode="session_userStoryMode"
@@ -27,7 +53,7 @@
 
     <div v-if="!planningStart">
       <div id="catGifDiv">
-        <b-img :src="require('@/assets/LoadingCat.gif')" class="catGif"/>
+        <b-img :src="require('@/assets/LoadingCat.gif')" class="catGif" />
       </div>
       <copy-session-id-popup
         :text-before-session-i-d="$t('page.session.before.text.beforeID')"
@@ -35,7 +61,6 @@
         :text-after-session-i-d="$t('page.session.before.text.afterID')"
         class="copy-popup"
       />
-
 
       <b-row class="d-flex justify-content-center overflow-auto kick-user">
         <kick-user-wrapper
@@ -48,28 +73,65 @@
       </b-row>
       <b-row>
         <b-col class="text-center">
-          <session-start-button @clicked="onPlanningStarted" />
+          <session-start-button
+            :members="members"
+            :host-voting="session_hostVoting"
+            :auto-reveal="autoReveal"
+            @clicked="onPlanningStarted"
+          />
         </b-col>
       </b-row>
     </div>
-
     <div v-else>
       <b-row class="d-flex justify-content-start pb-3">
         <b-col cols="auto" class="mr-auto optionButtonCol">
           <b-button
             class="mr-3 optionButton"
             variant="outline-dark"
-            @click="sendRestartMessage(); $event.target.blur();"
+            @click="
+              sendRestartMessage();
+              $event.target.blur();
+            "
           >
             <BIconArrowClockwise class="bIcons"></BIconArrowClockwise>
             {{ $t("page.session.during.estimation.buttons.new") }}
           </b-button>
-          <b-button class="mr-3 optionButton"
+          <b-button
+            class="mr-3 optionButton"
             variant="outline-dark"
-            @click="sendVotingFinishedMessage(); $event.target.blur();"
+            @click="
+              sendVotingFinishedMessage();
+              $event.target.blur();
+            "
           >
             <BIconBarChartFill class="bIcons"></BIconBarChartFill>
             {{ $t("page.session.during.estimation.buttons.result") }}
+          </b-button>
+          <b-button
+            v-if="!autoReveal"
+            class="mr-3 optionButton"
+            variant="outline-dark"
+            :disabled="planningStart && !estimateFinished"
+            @click="
+              autoReveal = true;
+              $event.target.blur();
+            "
+          >
+            <b-icon-eye-slash-fill class="bIcons" />
+            {{ $t("page.session.during.estimation.buttons.autoRevealOff") }}
+          </b-button>
+          <b-button
+            v-if="autoReveal"
+            class="mr-3 optionButton"
+            variant="outline-dark"
+            :disabled="planningStart && !estimateFinished"
+            @click="
+              autoReveal = false;
+              $event.target.blur();
+            "
+          >
+            <b-icon-eye-fill class="bIcons" />
+            {{ $t("page.session.during.estimation.buttons.autoRevealOn") }}
           </b-button>
         </b-col>
         <b-col cols="auto">
@@ -77,7 +139,7 @@
             :start-timestamp="timerTimestamp"
             :pause-timer="estimateFinished"
             :duration="timerCountdownNumber"
-            :votingStarted="planningStart"
+            :voting-started="planningStart"
             @timerFinished="sendVotingFinishedMessage"
           />
         </b-col>
@@ -98,14 +160,34 @@
         />
       </b-row>
       <hr class="my-5 breakingLine" />
-      <h4>
+      <h4 v-if="!session_hostVoting">
         {{ $t("page.session.during.estimation.message.finished") }}
         {{ membersEstimated.length }} /
-        {{ membersPending.length + membersEstimated.length }}
+        {{ members.length }}
+      </h4>
+      <h4 v-else>
+        <div v-if="hostEstimation == ''">
+          {{ $t("page.session.during.estimation.message.finished") }}
+          {{ membersEstimated.length }} /
+          {{ members.length + 1 }}
+        </div>
+        <div v-else>
+          {{ $t("page.session.during.estimation.message.finished") }}
+          {{ membersEstimated.length + 1 }} /
+          {{ members.length + 1 }}
+        </div>
       </h4>
       <b-row
+        v-if="highlightedMembers.includes(adminID)"
         class="my-1 d-flex justify-content-center flex-wrap overflow-auto kick-user"
+        style="max-height: 500px"
       >
+        <session-admin-card
+          v-if="(estimateFinished && session_hostVoting) || hostEstimation !== ''"
+          :current-estimation="hostEstimation"
+          :estimate-finished="estimateFinished"
+          :highlight="highlightedMembers.includes(adminID) || highlightedMembers.length === 0"
+        />
         <kick-user-wrapper
           v-for="member of estimateFinished ? members : membersEstimated"
           :key="member.memberID"
@@ -118,10 +200,51 @@
           }"
         />
       </b-row>
+      <b-row
+        v-else
+        class="my-1 d-flex justify-content-center flex-wrap overflow-auto"
+        style="max-height: 500px"
+      >
+        <kick-user-wrapper
+          v-for="member of estimateFinished ? members : membersEstimated"
+          :key="member.memberID"
+          child="SessionMemberCard"
+          :member="member"
+          :props="{
+            estimateFinished: estimateFinished,
+            highlight:
+              highlightedMembers.includes(member.memberID) || highlightedMembers.length === 0,
+          }"
+        />
+        <session-admin-card
+          v-if="(estimateFinished && session_hostVoting) || hostEstimation !== ''"
+          :current-estimation="hostEstimation"
+          :estimate-finished="estimateFinished"
+          :highlight="highlightedMembers.includes(adminID) || highlightedMembers.length === 0"
+        />
+      </b-row>
+      <div v-if="session_hostVoting && estimateFinished === false">
+        <div v-if="!estimateFinished">
+          <hr class="breakingLine" />
+          <h4 class="d-inline">Your Estimation</h4>
+        </div>
+        <div v-if="!estimateFinished" class="newVotes m-1">
+          <b-button
+            v-for="item in voteSet"
+            :key="item"
+            class="activePills m-1"
+            pill
+            style="width: 60px"
+            @click="vote(item)"
+          >
+            {{ item }}
+          </b-button>
+        </div>
+      </div>
     </div>
     <b-row v-if="session_userStoryMode !== 'NO_US'" class="mt-4">
       <b-col>
-        <user-story-sum-component/>
+        <user-story-sum-component />
       </b-col>
     </b-row>
     <b-row v-if="session_userStoryMode !== 'NO_US'">
@@ -129,7 +252,10 @@
         <div v-if="session_userStoryMode === 'US_JIRA'" class="refreshUserstories">
           <b-button
             class="w-100 mb-3 refreshButton"
-            @click="refreshUserStories(); $event.target.blur();"
+            @click="
+              refreshUserStories();
+              $event.target.blur();
+            "
           >
             {{ $t("page.session.before.refreshStories") }}
           </b-button>
@@ -148,7 +274,10 @@
         <div v-if="session_userStoryMode === 'US_JIRA'" class="refreshUserstories">
           <b-button
             class="w-100 mb-3 refreshButton"
-            @click="refreshUserStories(); $event.target.blur();"
+            @click="
+              refreshUserStories();
+              $event.target.blur();
+            "
           >
             {{ $t("page.session.before.refreshStories") }}
           </b-button>
@@ -205,6 +334,7 @@ import KickUserWrapper from "@/components/KickUserWrapper.vue";
 import SessionCloseButton from "@/components/actions/SessionCloseButton.vue";
 import SessionStartButton from "@/components/actions/SessionStartButton.vue";
 import { BIconArrowClockwise, BIconBarChartFill } from "bootstrap-vue";
+import SessionAdminCard from "@/components/SessionAdminCard.vue";
 
 export default Vue.extend({
   name: "SessionPage",
@@ -219,20 +349,22 @@ export default Vue.extend({
     UserStoryDescriptions,
     NotifyHostComponent,
     BIconArrowClockwise,
-    BIconBarChartFill
-},
+    BIconBarChartFill,
+    SessionAdminCard,
+  },
   props: {
-    adminID: { type: String, required: false },
-    sessionID: { type: String, required: false },
-    voteSetJson: { type: String, required: false },
-    sessionState: { type: String, required: false },
-    timerSecondsString: { type: String, required: false },
+    adminID: { type: String, required: false, default: undefined },
+    sessionID: { type: String, required: false, default: undefined },
+    voteSetJson: { type: String, required: false, default: undefined },
+    sessionState: { type: String, required: false, default: undefined },
+    timerSecondsString: { type: String, required: false, default: undefined },
+    hostVoting: { type: String, required: true },
     startNewSessionOnMountedString: {
       type: String,
       required: false,
       default: "false",
     },
-    userStoryMode: { type: String, required: false },
+    userStoryMode: { type: String, required: false, default: undefined },
     rejoined: { type: String, required: false, default: "true" },
   },
   data() {
@@ -244,6 +376,7 @@ export default Vue.extend({
       session_sessionState: "",
       session_timerSecondsString: "",
       session_userStoryMode: "",
+      session_hostVoting: false,
       //data
       index: 0,
       stageLabelReady: "Ready",
@@ -254,6 +387,8 @@ export default Vue.extend({
       startTimerOnComponentCreation: true,
       estimateFinished: false,
       session: {},
+      hostEstimation: "",
+      autoReveal: false,
     };
   },
   computed: {
@@ -320,8 +455,14 @@ export default Vue.extend({
       }
     },
     membersEstimated() {
-      if (this.membersPending.length  === 0 && this.membersEstimated.length > 0) {
-        this.estimateFinished = true;
+      if (this.membersPending.length === 0 && this.membersEstimated.length > 0 && this.autoReveal) {
+        if (this.session_hostVoting && this.hostEstimation !== "") {
+          console.log("HIER?????");
+          this.estimateFinished = true;
+        } else if (!this.session_hostVoting) {
+          console.log("hier?");
+          this.estimateFinished = true;
+        }
       }
     },
   },
@@ -350,6 +491,7 @@ export default Vue.extend({
       this.planningStart = true;
     } else if (this.session_sessionState === Constants.memberUpdateCommandVotingFinished) {
       this.planningStart = true;
+      console.log("ON MOUNTED");
       this.estimateFinished = true;
     }
   },
@@ -402,6 +544,7 @@ export default Vue.extend({
         this.session_timerSecondsString = this.timerSecondsString;
         this.session_voteSetJson = this.voteSetJson;
         this.session_userStoryMode = this.userStoryMode;
+        this.session_hostVoting = String(this.hostVoting).toLowerCase() === "true";
       }
     },
     assignSessionToData(session) {
@@ -412,6 +555,7 @@ export default Vue.extend({
         this.session_timerSecondsString = session.sessionConfig.timerSeconds.toString();
         this.session_voteSetJson = JSON.stringify(session.sessionConfig.set);
         this.session_userStoryMode = session.sessionConfig.userStoryMode;
+        this.session_hostVoting = String(session.hostVoting).toLowerCase() === "true";
         this.$store.commit("setUserStories", {
           stories: session.sessionConfig.userStories,
         });
@@ -426,6 +570,7 @@ export default Vue.extend({
         this.planningStart = true;
       }
       if (this.session_sessionState === Constants.memberUpdateCommandVotingFinished) {
+        console.log("ON HANDLE RELOAD");
         this.estimateFinished = true;
       }
       this.timerCountdownNumber = parseInt(this.session_timerSecondsString, 10);
@@ -444,7 +589,11 @@ export default Vue.extend({
       if (this.session_userStoryMode === "US_JIRA") {
         let response;
         if (doRemove) {
-          response = await apiService.deleteUserStory(us[idx].id);
+          if (us[idx].id === null) {
+            response = 204;
+          } else {
+            response = await apiService.deleteUserStory(us[idx].id);
+          }
           us.splice(idx, 1);
           doRemove = false;
         } else {
@@ -467,7 +616,11 @@ export default Vue.extend({
           }
         }
         if (response.status === 200) {
-          this.$toast.success(this.$t("session.notification.messages.issueTrackerSynchronizeSuccess"));
+          this.$toast.success(
+            this.$t("session.notification.messages.issueTrackerSynchronizeSuccess")
+          );
+        } else if (response === 204) {
+          this.$toast.info(this.$t("session.notification.messages.issueTrackerNothingChanged"));
         } else {
           this.$toast.error(this.$t("session.notification.messages.issueTrackerSynchronizeFailed"));
         }
@@ -519,7 +672,9 @@ export default Vue.extend({
           }
         }
         if (response.status === 200) {
-          this.$toast.success(this.$t("session.notification.messages.issueTrackerSynchronizeSuccess"));
+          this.$toast.success(
+            this.$t("session.notification.messages.issueTrackerSynchronizeSuccess")
+          );
         } else {
           this.$toast.error(this.$t("session.notification.messages.issueTrackerSynchronizeFailed"));
         }
@@ -560,15 +715,23 @@ export default Vue.extend({
     },
     sendVotingFinishedMessage() {
       if (!this.estimateFinished) {
-        const endPoint = Constants.webSocketVotingFinishedRoute;
-        this.$store.commit("sendViaBackendWS", { endPoint });
+        console.log("SENDVOTINGFINISHEDMESSAGE");
         this.estimateFinished = true;
+        const endPoint = `${Constants.webSocketVotingFinishedRoute}`;
+        this.$store.commit("sendViaBackendWS", { endPoint });
       }
     },
     sendRestartMessage() {
       this.estimateFinished = false;
+      this.hostEstimation = "";
       const endPoint = Constants.webSocketRestartPlanningRoute;
-      this.$store.commit("sendViaBackendWS", { endPoint });
+      this.$store.commit("sendViaBackendWS", {
+        endPoint,
+        data: JSON.stringify({
+          hostVoting: this.session_hostVoting,
+          autoReveal: this.autoReveal,
+        }),
+      });
     },
     goToLandingPage() {
       this.$router.push({ name: "LandingPage" });
@@ -576,11 +739,34 @@ export default Vue.extend({
     onPlanningStarted() {
       this.planningStart = true;
     },
+    vote(vote: string) {
+      this.hostEstimation = vote;
+      const endPoint = `${Constants.webSocketVoteRouteAdmin}`;
+      this.$store.commit("sendViaBackendWS", {
+        endPoint,
+        data: JSON.stringify({
+          vote: this.hostEstimation,
+          autoReveal: this.autoReveal,
+        }),
+      });
+    },
   },
 });
 </script>
 
+<!-- Add "scoped" attribute to limit CSS to this component only -->
 <style scoped>
+.newVotes {
+  text-align: center;
+  margin-left: auto;
+  margin-right: auto;
+}
+
+.hostVotingButtons {
+  position: relative;
+  top: 50%;
+  transform: translateY(-50%);
+}
 
 .optionButtonCol {
   margin-top: auto;
@@ -614,38 +800,42 @@ export default Vue.extend({
   width: 40px;
 }
 
-.optionButton{
+.optionButton {
   background-color: var(--textAreaColour);
   color: var(--text-primary-color);
   border-color: black;
-  border-radius: var(--buttonShape);
   display: inline-flex;
   align-items: center;
 }
 
-.optionButton:hover{
+.optionButton:hover {
   background-color: var(--textAreaColourHovered);
   color: var(--text-primary-color);
 }
 
-.optionButton:focus{
+.optionButton:focus {
   background-color: var(--textAreaColourHovered) !important;
-  color: var(--text-primary-color)  !important;
+  color: var(--text-primary-color) !important;
+}
+
+.optionButton:disabled {
+  background-color: var(--textAreaColourHovered) !important;
+  color: var(--text-primary-color) !important;
 }
 
 .refreshButton {
   border-radius: var(--element-size);
   color: var(--text-primary-color);
-  background-color: var(--joinButton);
+  background-color: var(--secondary-button);
 }
 
 .refreshButton:hover {
   color: var(--text-primary-color);
-  background-color: var(--joinButtonHovered);
+  background-color: var(--secondary-button-hovered);
 }
 
 .refreshButton:focus {
-  background-color: var(--joinButtonHovered) !important;
+  background-color: var(--secondary-button-hovered) !important;
   color: var(--text-primary-color) !important;
 }
 
@@ -656,5 +846,20 @@ export default Vue.extend({
 .catGif {
   width: 240px;
   height: 180px;
+}
+
+.activePills {
+  background-color: var(--preparePageMainColor);
+  color: var(--text-primary-color);
+}
+
+.activePills:hover {
+  background-color: var(--preparePageInActiveTabHover);
+  color: var(--text-primary-color);
+}
+
+.activePills:focus {
+  background-color: var(--preparePageInActiveTabHover) !important;
+  color: var(--text-primary-color);
 }
 </style>
